Update patch set 1
Patch Set 1: Code-Review-1

As I understand it committers are appointed on an individual basis (not company). To add a committer requires a majority vote from exiting committers. That is the procedure I followed to add committers to functest and included links to the voting results in the file (request by Chris Price) ... https://gerrit.opnfv.org/gerrit/#/c/508/1//COMMIT_MSG I think we need to clarify if in fact "swapping" committers is allowed before merging this change. Of course contributors are always easy to accommodate :)

Patch-set: 1
Label: Code-Review=-1