Update patch set 1
Patch Set 1: Code-Review-1

Please hold on with this since we still haven't got any outcome from scenario discussions and whatever you see on Jenkins is my own speculation which is probably wrong.

I had a discussion with Jonas and we will probably have something for fuel tomorrow and the Jenkins jobs will be updated according to this.

Once we have that, we have to convince the rest to follow it since we don't have luxury to discuss things same way as we have been doing and stick to what Jonas comes up with.

Sorry for the inconvenience.

Patch-set: 1
Reviewer: Gerrit User 254 <254@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=-1