Update patch set 3
Patch Set 3: Code-Review-1

(2 comments)

will not work due to the hardcoded env variable in doctor.py

excepct if you stick to one installer?

any constraint to run doctor?
if it is on fuel only then it could maybe work (I have to check the default value of the env var) but it is recommended anyway to use available vars

Patch-set: 3
Label: Code-Review=-1