Update patch set 1
Patch Set 1: Code-Review-1

As the file is for the moment mainly a translation of the bash (which is a good first step) but does not leverage the Functest frameworks (reuse of config file, dashboard, ..) I put -1 to let you time to update this script. 
It will be difficult to use it for Brahmaputra SR0 but we could use it for Brahmaputra SR1.
This week we must finalize the documents and make sure the tests can be run.
We will have also integration tasks regarding projects that just arrived
We will have more time to help you on the refactoring after the rush of B Release

Patch-set: 1
Reviewer: Gerrit User 335 <335@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=-1