Option to redirect stderr output to a file for Tempest test
The output when we run Tempest from CI is too long
if there are many tests failing. This patch intends to
show only the test cases that passed/failed within a table.
If troubleshooting is needed, it will be a log file with all
the stderr in it.

Change-Id: If5b0f35463a3dae06a0c177de4dc473bfc4687c5
Signed-off-by: jose.lausuch <jose.lausuch@ericsson.com>
(cherry picked from commit 9be6519eca39111295b89e9b3b0d853f95126a72)