Reduce Rally output
This patch intends to reduce the output of Rally test,
showing only the important information of the tests,
which is the scenario name and the table with results.
Showing the scenario json/yaml is unnecessary since it is
already in the repo.

Change-Id: Iff0ba3fed17212bb4a695c4155747c8aaf683da3
Signed-off-by: jose.lausuch <jose.lausuch@ericsson.com>
(cherry picked from commit b0365d10fda5018721d012602077dddda948a6d9)