DO NOT MERGE:  Udating structure and heading levels
I have decomposed the config and user files into some sub-file
elements.  The sub-file elements will be pulled into the composite
docs to fill out sections of the user and config guides.
This is the draft version to check I have the basics right, comment please
but do not merge at this time.

Change-Id: If46152db14aa1becc37e63e4095981d44e7e34ef
Signed-off-by: ChristopherPrice <christopher.price@ericsson.com>