Update patch set 2
Patch Set 2:

Hi,  please comment on the errors if we fix them now it's lee cherry picking.  The reason to seperate the docs is to have sub-files that will be pulled in the "brahmaputra config" and "brahmaputra user" guides about our functest project.  We should see how it looks once merged and make adjustments as neccesary.

Patch-set: 2