Add script to generate the OpenStack defaults
JIRA: FUNCTEST-131

It checks the existing Openstack objects in the current
deployment (i.e. when the script is executed) and it will
create a yaml file to be used afterwards by clean_openstack.

This way, a user can generate the defaults when needed if there
are things that are not meant to be deleted without fear of
using clean_openstack.

Change-Id: Iacf204bdd14f591df21b3f243d27199ead2404b0
Signed-off-by: jose.lausuch <jose.lausuch@ericsson.com>
(cherry picked from commit 954907ac2995cfc158d5e363829af7c776be5f90)