Adding a structure for the reporting of test cases for functest
The structure builds on an index.rst in the results folder.
The results are seperated inot specific files for each scenario.
All scenarios are commented out at this time and we can add them when ready.
Not all scenario files have template text at this time.  (have to run, wanted to push this)

Change-Id: Iaff140a5e8106f19d15d23f937f6dd27453a5680
Signed-off-by: ChristopherPrice <christopher.price@ericsson.com>