Update patch set 1
Patch Set 1: Code-Review-1

That's is right and although this might be a temporary solution, we shouldn't go for installer-specific test cases. As Viktor said, the hard coded list is going to disappear since it's difficult to maintain. Tempest will generate automatically the needed test cases on the fly according to the arguments (smoke, full, etc... ). That is what the OpenStack community also advises. 
So, if you don't mind I give a -1 to this patch since this is gonna be solved soon in a different way already agreed by Functest members. Is it ok?

Patch-set: 1
Reviewer: Gerrit User 459 <459@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=-1