Update patch set 1
Patch Set 1:

In the Functest code now there are "suites" which are just the tempest test cases listed in json format.  I see there opnfv-smoke.json and opnfv-smoke-green.json.  Both of these list test cases out explicitly and I think would have the same problem as keeping a list in text format.  Can you please explain how these are different?

Also, I think installers need to have input on what tests are valid as smoke tests.  Things like IPV6 ports and router tests should not be in smoke tests.  The installers should get say in what gates their commits.

Does it make more sense if we keep the test case list in our repo so we control it, then we just copy it over the custom list when we invoke functest?  We could do that for now in our setup until we come to a better agreement on this.

Patch-set: 1