Update patch set 1
Patch Set 1:

opnfv-smoke.json and opnfv-smoke-green.json will not be used any more when we have the new way to run tempest. The "smoke" list will be auto-generated by tempest according to the deployment (tempest does this with some intelligence). You can keep it in your repo, but we will not run hard coded lists of test cases any more :)
We might have a blacklist of test cases in case you want to exclude some. But let

Patch-set: 1