Update patch set 2
Patch Set 2:

Bryan, it's completely fine if you want to keep copper.py in your repo. We don't mind at all. If you are planning to improve/maintain it in your repo, then we don't need to worry about keep it updated it in functest (we also like upstream). As you wish.

Patch-set: 2