Update patch set 1
Patch Set 1:

Yes, but we don't pull all the repos, only those that we need, since this is a new feature in Functest, it wasn't existing before :) Please have a look at the Dockerfile located in the docker directory and you'll see what I'm refering to. Update it adding a line to clone  the domino repo.

Patch-set: 1