Use blacklist of Tempest test cases for SDNVPN scenario
This patch disables tempest smoke from being executed in bgpvpn scenarios.
Since there is a known upstream issue which won't be resolved before Colorado,
it doesn't make sense to run the test cases in Tempest that will fail due
to that issue.
The Tempest suite will be triggered along with the "features" tier, with
a given blacklist of the known test that will fail.

JIRA: FUNCTEST-369

Change-Id: Ifc533077207e9d62d49bcedcc3cd0936df312891
Signed-off-by: jose.lausuch <jose.lausuch@ericsson.com>