Update patch set 6
Patch Set 6:

This solution is not scalable. What happens if more features want to change tempest.conf? We put the code here? This script would become unreadable and too long. And we shouldn't touch the original tempest.conf... 

I have created this JIRA to work on a common solution:
https://jira.opnfv.org/browse/FUNCTEST-377

Please take a look at how we have dont things in SDNVPN project:
https://gerrit.opnfv.org/gerrit/#/c/16895/1/test/functest/tempest.py

We create a copy of tempest.conf and run tempest directly from the command line specifying the path to the modified tempest.conf (not the best way but it works 100%)

I can merge this only if you promise to work on doing it the same way and removing this code from run_tempest.py

Patch-set: 6