Bump rally to 0.5.0 and add shyaml
A newer version of rally brings in this change that is required by armband:
https://bugs.launchpad.net/rally/+bug/1580695

The shyaml cli parser can be used to make the healthcheck testcase
configurable, for instance to pass a different cirros image other than the one
for x86_64. Also required by armband to make functest execute all tests.

Change-Id: Ic4976bb6d4a4753e56deade35b641f1127af3162
Signed-off-by: Ciprian Barbu <ciprian.barbu@enea.com>