modfied to mute object storage test on joid.
In case of joid we implement ceph-radosgw for object storage which creates
swift api for object storage. Manually i can verify we can upload the
object. But functest is failing other installer removed the API but we
need in JOID for juju based deployment on Openstack. So this patch will
mute the object storage test cases for functest.

Change-Id: I21d2bdc665a9589630cfcacb278cfb65a1c15f11
Signed-off-by: Narinder Gupta <narinder.gupta@canonical.com>