Avoid duplicating logs in run_tests.py
When run_tests.py sets logger as execute_command arg, logs are
duplicated in functest.log and not shown in console (as every DEBUG
message). As logger is mainly managed by the python scripts in charge
of testing, it can be safely removed here.
stderr is also redirected to stdout in execute_command to print
possible relevant data.

JIRA: FUNCTEST-431

Change-Id: Ie77544b4679e12e0a0ac7f5e5989d86862fe106b
Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>