Committer promotion: Cedric Ollivier
Cedric joind Functest for Colorado
He fully refactored the ODL section including work upstream within ODL community
In addition of ODL skills, Cedric is a great benevolent reviewer.
Even if he is famous in the project for his gerrit -1, his reviews are always
usefull and clearly consolidate the code baseline for functest.
He is very relevant on all the technical aspects and we always need a Debianist
among commiters

Change-Id: I4bd0f778edfaa0d8969bdda3387d5a5ebc8b0f93
Signed-off-by: Morgan Richomme <morgan.richomme@orange.com>