Update patch set 2
Patch Set 2: Code-Review+2

> why the special process for ODL?

I think it should be generalized in D (part of the refactoring with template/object view). the idea is to avoid python => bash => python, which is an (uggly) historical heritage. Cedric did the patch for ODL as he is taking care of ODL. For D we should have some abstraction for any case.
So for C it may look a bit strange to consider odl on one side and the other cases on the other side, we may consider as a first try.
As we need an abstraction a generalization for C is not trivial and too late regarding the milestones.

my only concern was the consequences on CLI but in CLI we call run_tests.py, so no problem.

Patch-set: 2
Label: Code-Review=+2