Update patch set 2
Patch Set 2:

> So, some work in Release D can be proceeded now, Can I work on
 > FUNCTEST-332 now?

we can always work for future release in master :)...however we cherry pick only fixes, "safe" improvements (this patch could be considered in this category) or release related document changes.
the priority is then to finalize the Colorado release (still some work on dashboarding, no?) and prepare the future...but for the FUNCTEST-332, I think we have to discuss it as for D we planned the integration of SNAPS (the cable labs testing framework), which is already object oriented...we need to agree on the architecture first before starting coding. We may open an etherpad and initiate a wiki page (I add this point for next week weekly release: D release....)

Patch-set: 2