(my idea)Create Template for feature project integration
Because in functest all the featuer project have similar process.
The major two differences is the 'cmd' and 'criteria' check.
As to 'cmd', it depends on repo and execute file, so we can abstract
this different using 'cmd' to represent it in testcases.yaml
And as to 'criteria' check, we can abstract 'criteria' to represent
which criteria validity method is used to check the result instead of
status/success_rate.

As to CI execute, Use run_test.py to substitue exec_test.sh

After doing this all the feature project repo in testcases
whoes actual test work not done in functest can be deleted(just as I did to parser),
Only testcases.yaml need to be maintained.

As to vping/tempest/rally/odl...,we can also abstract the
actual test work and the driven work by functest, and apply
the driven work to the framework.

This is a brief version, just to indicate my idea

Change-Id: Id9a90be93be489a136c1b839a01b21a3e8f3be87
Signed-off-by: SerenaFeng <feng.xiaowei@zte.com.cn>