Update patch set 3
Patch Set 3:

This will be used for test cases that print messy output to Jenkins, specially those which don't use functest logger mechanism. A log file must be generated and pushed to artifacts instead of showing the output in Jenkins. This will help to maintain clean our job execution.

Patch-set: 3