[Promise test case] Use proper way to get parameters from yaml
Since there is a function in functest_utils to get
a value from the config_functest.yaml directly, there
is no need to load again that file. Better to use common
functions that are designed for that.

Also, Replace location of the results json file to
the results directory. It will be also pushed to artifacts
repository.

Change-Id: Ic6422ed78ccd0f8b40b519342a7dcf38c502924e
Signed-off-by: jose.lausuch <jose.lausuch@ericsson.com>