Add support of arm for doctor
This is half of the fix to make the doctor suite run on arm.
The other half is at https://gerrit.opnfv.org/gerrit/#/c/20369/
In this fix I set an enviroment variable, to be used by
the doctor suite when ran with functest. The variable is set only
when present in config_functest.yaml, which we use as a custom file
when running functest on arm.

Change-Id: If34d3e7734a194bf9fe8b525a362e029767caa4f
Signed-off-by: cristinapauna <cristina.pauna@enea.com>
(cherry picked from commit 53b8060cd60c163315d706a6035e30eb66fe8d15)