Update patch set 2
Patch Set 2:

We can keep a wrapper for calling shell scripts, as we do for copper.py, promise.py, doctor.py, etc. They are very similar and in the end they call an external whatever script (normally bash). I wouldn't like to impose external test to be written in the language we want. We should be flexible and adapt to it in a clever way.

Patch-set: 2
Reviewer: Gerrit User 459 <459@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=0