Update patch set 2
Patch Set 2:

Hi Morgan, I also have a question. In devguide/index.rst, lines 146-147 and lines 260-261, I see the link of testresuls and the authentication are listed in two lines. In the meantime, in the html version, they are in one line. Is this a bug for document build script?

Patch-set: 2