[SDNVPN] Disable Rally Sanity test case for bgpvpn scenarios
Due to a known issue with API performance when bgpvpn extension
is activated which causes some delays and the tests take longer
time than normal, we get a lot of Jenkins Job timeouts and the
scenario cannot be validated.
Since it is a performance issue, we black list rally sanity, which
takes 1 hour to complete and will re-enable it in Colorado.2.0
where ODL Boron release will be available and SDNVPN team
thinks this will be fixed.

Change-Id: Idc68e30ea15d52d0fd4b63bc57ddf643f3f7c8b4
Signed-off-by: jose.lausuch <jose.lausuch@ericsson.com>