Update patch set 1
Patch Set 1:

Thank for reporting the difference. 
Let see what Morgan and devs think. Do we continue this review as-is or do we switch it to fix the other testcases?
I think if the user asks to report to db and it fails, we should exit the corresponding value.

Patch-set: 1