[ODL-SFC] Open port 22 in default secgroup for sfc scenario
JIRA: SFC-48

After some troublehsooting done in Boron RC3.5, vPing
doesn't work because the instances are created on 2
security groups. One (vping-sg) opens port 22 but
the default secgroup doesn't. This patch allows port
22 also on the default and therefore vPing passes.

This patch is to workaround a bug in ODL. The bug
will be reported to the corresponding community.

Change-Id: I807249d42cebe7c8dccd78a3496c1d34996d73ee
Signed-off-by: jose.lausuch <jose.lausuch@ericsson.com>