Move sfc custom flavor creation
The custom flavor that is needed for the sfc test case is created in the
tacker deployment phase.
The result of this is that when we use alternative methods of deploying
tacker (i.e. the tacker fuel plugin) the test fails.

Changes:
- Move the flavor creation inside the test case code
- Add an optional public parameter in the [get_or_]create_flavor utility
  functions.

Change-Id: Iea13f172076770dde3876b7e27edc9b7ca9b28db
Signed-off-by: George Paraskevopoulos <geopar@intracom-telecom.com>
(cherry picked from commit 40f68e35389cf19d3fb200d1655ecb055c85d1eb)