Update patch set 1
Patch Set 1:

I was pondering that... basically just to separate full/sanity and to make clearer that a modification alloted to one scenario only won't affect the other one. This of course can be a drawback too if a change is such where a common change is desirable.

One option would be to divide only in those scenarios where the content actually differ. Rest of the scenarios continue to be shared ones until there's a real need for division. What do you think about this?

Patch-set: 1