Update patch set 1
Patch Set 1: Code-Review-1

set-up-tacker.sh is doing two important things which are still valid. Creating the delete.sh which is used in the "iterative testing" and creating the flavor for the SFs which we create through tacker. We can do these things in another way but then we should add that, otherwise we will break the test

Patch-set: 1
Label: Code-Review=-1