Update patch set 2
Patch Set 2:

In practice, is rally_conf.json the only file which is ignored here? If yes, would it make sense to add the deletion now instead of ignoring json categorically? E.g in case of rally, we are currently using test cases defined in yaml but rally supports json format too.

Patch-set: 2