Update patch set 2
Patch Set 2:

We have to ignore all volatile files which are generated by testcases. I prefer the generic way as we are currently using yaml files but if you disagree I can revert to the first review.
 
I don't bet about the future. Bu if we switch to json we will have to ignore yaml files as well.

Patch-set: 2