Update patch set 7
Patch Set 7:

> (2 comments)

agree but the way we play with env variable in the legacy code is sometimes a bit dangerous

I tried to setup the env var in the unit tests without success for the moment. I got an error in the import before I can run setup. I can obviously create the env var in a run_test.sh calling everything but I cannot do it by modifying "just" the unit test

Patch-set: 7
Label: Verified=0