Update patch set 3
Patch Set 3:

Based on what I saw in my local env, when we assign security group to existing vm this flows are not created:

    cookie=0x0, duration=16.700s, table=40, n_packets=0, n_bytes=0, priority=61007,tcp,dl_src=fa:16:3e:35:6e:d6,tp_dst=22,tcp_flags=syn actions=goto_table:50
    cookie=0x0, duration=16.709s, table=90, n_packets=0, n_bytes=0, priority=61007,tcp,dl_dst=fa:16:3e:35:6e:d6,tp_dst=22,tcp_flags=syn actions=goto_table:100

with proposed changes those flows are created. It can be something specific to this lab, so it would be good if someone could also verify this. If it will be confirmed I guess that it should be submitted as bug in ODL.

Patch-set: 3