Update patch set 3
Patch Set 3:

Adding an already created VM to a security group is valid use case in OpenStack. This patch is a workaround to make the test work. The way I see this is a fault in ODL and it should be filed as a bug in the community before we merge this. I guess this happens only in Boron, since this test was working in Colorado 1.0.

Patch-set: 3