Update patch set 3
Patch Set 3: Code-Review-1

I'm not sure about merging this, because this is not a good habit for collaboration between communities. We should use this as feedback so we get faster a fix. If we workaround the issue instead of pushing the upstream community to fix it, we might end up with a lot of hidden possible bugs that are not fixed since there are workarounds to show blue in jenkins.. The other reason I have to not workaround this issue is because it's a regression. It was working before and it doesn't work now. If it was something like a missing feature implementation (like floating ips in bgpvpn scenarios), we could merge it, but not the case.

Patch-set: 3
Reviewer: Gerrit User 459 <459@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=-1