Update patch set 3
Patch Set 3: Code-Review-1

(5 comments)

Nice work. I have some objections (a few of them are opinionated) but the testcase needed some refactoring. Also we need to think about what is going to happen with sfc_colorado1 testcase. Now that they have submitted the necessary bug fixes in ODL I think it can be phased out so that we don't have to do 2 times the work to refactor and maintain it (or even worse have it in a different state).

Patch-set: 3
Label: Code-Review=-1