Make the healthcheck sleep time configurable
The healthcheck test has a 1 min sleep that gives time to the odl
flows to get programmed. This time depends on the config, and it
was noticed that on bgpvpn scenarios on arm more than one minute
is needed for the config to be properly set.

This patch makes the waiting time configurable, so it can be
overriden it when ran on arm or on varios scenarios.

Change-Id: Iebef892137ca3926e5732c910a358c15a32b5bef
Signed-off-by: cristinapauna <cristina.pauna@enea.com>