Remove kingbird installation
This is a quick hack to make docker build work in CI.
Kingbird is used by multisite, but the test case
is empty. The installation of the requirements overlap
with what we already install in requirements.pip
and makes the build fail.
It can be enabled once we resolve the dependencies
across different projects.

Change-Id: I9194a7ae7a5887ee80eb392890dd304837760b66
Signed-off-by: jose.lausuch <jose.lausuch@ericsson.com>