Update patch set 5
Patch Set 5:

> (18 comments)
 > 
 > Morgan, I went through the code and most of the problems are
 > related to path. I would suggest to put all path related info into
 > a single file, say Constants or put all of them in functest_utils?
 > Currently, some values are stored in functest_utils, some are
 > passed from environment parameter, some are calculated by script,
 > which will cause unexpected error. Moreover, I would suggest to use
 > relative path instead of absolute path in some cases.

I agree with your comments, you can create a dedicate pathc for that.
This patch wasonly to restructure the repo. No change on the path management or renaming of the files to be homogeneous. Just change of structure. Docker file has been produced. We will see on the next master run if we did not break things. 
But clealery, your patch will be very useful

Patch-set: 5