Update patch set 1
Patch Set 1:

This patch adds in a new tier of tests called snaps that contains 2 new test suites. These could be added under healthcheck but they don't relate to each other in terms of runtime or configuration. Please see the attached PDF on the Jira ticket for an explanation of it's tests.

Patch-set: 1