Update patch set 1
Patch Set 1:

(1 comment)

First of all, I am sorry that I haven't tested it yet, so I don't add reviewers.
I think push2db is implemented in TestCasesBase, since there's no special requirements in vping, there's no need to override it, if it not the case, I would like to refactor it to make it work in this way.
With  respect to the format, I think it makes sense to follow the OpenStack, '_' to the module/file name and UpperCase to the object. 
Regarding to unit tests, I think the intern will do it, no?

Patch-set: 1