Update patch set 4
Patch Set 4: Code-Review+1

Looks good. Thanks for removing the parse of the criteria in the yaml. I think it should be 0 or !=0    .

I agree with Morgan's comment about the 

EX_TEST_FAIL = os.EX_SOFTWARE - 3

To me it is similar to EX_RUN_ERROR = os.EX_SOFTWARE

Patch-set: 4
Label: Code-Review=+1