Update patch set 8
Patch Set 8:

I see no glaring issues; however as we are trying to migrate these type of operations to SNAPS, should we consider making changes to that library instead? I would like to begin efforts to start a SNAPS (or whatever name) project under the OPNFV umbrella and will be discussing the logistics with Jose in NH next week.

Patch-set: 8