Update patch set 2
Patch Set 2: Code-Review+1

Jose, mapping in the local directory is a hack and does not work 100% of the time. Additionally, unless the changes have already been merged, you will never be able to create an image that is identical to the one that is built after the code has been merged. Please look closely at the reasons put forth within FUNCTEST-577. We can discuss in greater detail next week.

Patch-set: 2