Update patch set 8
Patch Set 8:

that is very valid point. The reason of refactoring our openstack_utils module was to have a consolidated way of calling the different APIs. If we can't use python-openstackclient, maybe we should consider SNAPs (or whatever name). 

Steven, can you give us a "workshop" about how to use the library and the different methods? That would be very useful for us and appreciated.

Patch-set: 8