Unified way to provide configurations and env variables(proposal 2)
In this proposal the env part keep the same with previous version.

As regards to configurations, the naming in config_functest.yaml can be arbitrary
and keep the current status, as we will rename them in config.py.
The disadvantages is that we need to sync config.py with yaml,
which is the dislike thing by Jose and most of us I guess.

Change-Id: I94340b9ea89399e843b6050331f582d3391471ca
Signed-off-by: SerenaFeng <feng.xiaowei@zte.com.cn>