Refactoring of args and parser variable in ci/run_tests.py
Stackoverflow link:
http://stackoverflow.com/questions/24397258/
can-not-run-nosetests-when-i-use-argparse-in-my-python-code

Can't import run_tests in test_run_tests.py (new unit test file).
It is throwing error: nosetests: error: unrecognized arguments

By looking at SO, I found this is because of declaration of
parser and args which are global in file, So to allow safe
import of this file, we need to declare it into __main__ branch.

Change-Id: Ie11ebcfd8a1b8e692bd8bf4260a54f752b67fd5e
Signed-off-by: ashishk1994 <ashishk.iiit@gmail.com>