Update patch set 2
Patch Set 2: Code-Review+2

same idea than 26657
@Serena I think I understood your concern
I would say that for unit tests we should focus on the code we want to test. The code under test must be of course optimized (and then factorized when possible). The mock objects are there to test the code and there is no need to factorize the mock themselves. 
my concern is that my mock knowledge is not good enough (and additionnaly that I am a poor coder) but I like the idea to do it this way...
Of course each unit test will have to implement similar things, it simply means that there are similar processing but the focus is to test and optimize the code not to optimize the code used to optimize the code...not sure I am clear :)

Somehow it is connected to the morning discussions on the unit tests
But I think mock objects are an elegant way  to focus on the useful part of the code without creating artificial resources (files, logger) which are created in an end 2 end configuration but not needed for unit tests
maybe this can be part of the slide deck mentioned this morning regarding unit test strategy

Patch-set: 2
Label: Code-Review=+2