Update patch set 13
Patch Set 13:

(2 comments)

Serena, thanks for your feedback.
In regard to your point that the patch is making it complex, I don't agree with you. Putting the logging configuration into file is a widely used way for better logging manipulation. It's decouple the logging configure from the code which makes the code every easy to adopt different logging strategy.

Patch-set: 13