Update patch set 13
Patch Set 13:

Forgive me for the typo in last comment. Put the main idea here another time :)
In regard to your point that the patch is making it complex, I don't agree with you. Putting the logging configuration into a file is a widely used way for better logging manipulation. It decouples the logging configuring from the code which makes the code very easy to adopt different logging strategy.

Patch-set: 13