Update patch set 1
Patch Set 1: Code-Review-1

sorry, I will give -1 temporarily, because I don't know the background and side-effect. after we discuss it through, I will discard it. 
My question is after this patch, will libvirt driver always try to find large pages for guest RAM, or  will it try in dedicated scenario? 
will it affect the execution of other scenarios? is it better if we make it work in a dedicated scenario ovs as you mentioned?

Patch-set: 1