Update patch set 6
Patch Set 6:

> Would it be better to use some predefined untagged version (commit)
 > instead of using whatever happens to be the latest at the time of
 > cloning?

 > Would it be better to use some predefined untagged version (commit)
 > instead of using whatever happens to be the latest at the time of
 > cloning?

I totally understand your concern. How about we wait and see if the rally is breaking our build? 
I would say running our test against latest rally will bring a benefit that we can know the integration issue in an early phase. 
If we hold and wait for the release of 0.8.0, we will have a tough time handling with the integration issue(if there is), which will put us in risk for the D release.
Under the current circumstance, we are afford to fix the problem.

Patch-set: 6