Update patch set 5
Patch Set 5:

> +1 to that. Will you have time to work on that or do you want me to
 > take it (don't have much though)?

so, what's the consensus here? keep this patch and then we need to
get some triggers somewhere to call the suites? seems to me whatever
was calling ODLTests before is now going to be calling the netvirt
connectivity suites. but, I think we didn't want that?

Patch-set: 5
Label: Verified=0