Update patch set 17
Patch Set 17: Code-Review+1

Are there any issues with this patch?  

It seems like Steve has addressed the current set of comments.  I understand that the plan is to use this for Danube.  It would be nice to start using this test ahead of the release and not introduce it later in the release cycle.

Patch-set: 17