Update patch set 17
Patch Set 17:

> Are there any issues with this patch?
 > 
 > It seems like Steve has addressed the current set of comments.  I
 > understand that the plan is to use this for Danube.  It would be
 > nice to start using this test ahead of the release and not
 > introduce it later in the release cycle.

my view: no objection from my side, but I would like to see the smoke SNAPS tests PASS with different installers.
for the moment it is OK only with Apex, we got error with fuel
so there is a risk that vPing that is currently working with all the installers could be OK only in apex.
once we figured out what is the problem with the other isntallers, we could merge

Patch-set: 17