Add flavor creation in functest env prepare
So far, healthcheck.sh was creating m1.tiny flavor.
We are planning to remove that script and replace it by
SNAPS api and connection check.

Althought, a better approach to this would be that all
the tests create the flavor if needed, but these are the
tests that rely on this flavor:
 - vping ssh
 - vping userdata
 - rally
 - tempest
 - onos_sfc

Change-Id: I537f68a55d733b24b6b15f5cf3710bca40fbc622
Signed-off-by: jose.lausuch <jose.lausuch@ericsson.com>