Update patch set 3
Patch Set 3:

> Why are we generating RSA keys? I am bit lost here why these
 > changes are being recommended? Also as said earlier, this is the
 > wrong repository.

When running the security scan code, it checked the existence of rsa key. This is the main reason why the key is generated.

The logic in the patch is mainly about calling the script in securityscanning project. I should update the commit message to avoid the confusion.

Patch-set: 3