Proposal 1: Set up the logger whenever is needed
This way is trying to set up the logger whenever there is no existing
handler configured for logger.

Benefit:
Only limited set_up is called whenever is really needed

Shortage:
This way is spreading the set_up over the project which is not good for
scalability

Ideally, the logger is only needed to be configured once at the
entrypoint of the APP. However, the way Functest is implemented does not
share the context among modules.
1. prepare_env and run_test are triggered separately
2. the way to trigger testcase also does not share context with the
former two
3. functest_utils is imported before the set up for logger in
testacase

Change-Id: I5c287c4dc627cab56707525700e9bc45f354ffb4
Signed-off-by: helenyao <yaohelan@huawei.com>