Update patch set 6
Patch Set 6: -Code-Review

You're right.  But to pass the build for now, can we just place there a dummy (place-holder) file?  ...Once all these declaration files are included in the Functest container it makes it easier for us to do the internal testing prior to uploading the true barometer_collectd.py and all other files in the repo.

Patch-set: 6