Modifies check_os.sh to check services and cinder is optional
This patch adds being able to detect if the service is enabled on the
target openstack system prior to checking if the corresponding openstack
command for that service works.  It also makes cinder optional, and
warns if that service is not up.  We don't need cinder for every
deployment type, for example: CSIT.

Change-Id: I9556acabbd16f3b13f74b496812a5742b9000153
Signed-off-by: Tim Rozet <trozet@redhat.com>