Restore Kingbird functest
Kingbird functest was excluded from Functest due to python client
dependency conflict. But in Functest container, the kingbird was
installed only for tempest plugin descovery and execution of tempest
test cases. It's safe to restore kingbird installation in the Dockerfile
now.

The kingbird installation works after the functest container image
was built.

Change-Id: I44593dcf47dd1deed2f6a94b7e801230bae9b251
Signed-off-by: joehuang <joehuang@huawei.com>