Restore Kingbird functest
Kingbird functest was excluded from Functest due to python client
dependency conflict. But in Functest container, the kingbird was
installed only for tempest plugin descovery and execution of tempest
test cases. Kingbird is not sensitive to the python client version
of OpenStack. So it's safe to replace kingbird's client's requirements
to Functest's requirements. After the replacement, all client's
requirements will be kept in consistent after kingbird's installation.

The added script works after the functest container image was built.

Change-Id: I44593dcf47dd1deed2f6a94b7e801230bae9b251
Signed-off-by: joehuang <joehuang@huawei.com>