Add Tempest Refstack test case
Refstack is the OpenStack project that provides the tooling
to run the approved specific tempest test cases by the
Interop working group and by the OpenStack board
for OpenStack certification.

This adds the possibility for a Functest user to run this list
of tempest test cases by just issuing the command:
    functest testcase run tempest_refstack

The test is set in the weekly tier (category) which is
meant for weekly loops (not run in CI in Danube) and it
is only active when having a plan OpenStack deployment
(nosdn-nofeature).

Change-Id: I920b2a44cff418116008fdf514425e69dc959412
Signed-off-by: jose.lausuch <jose.lausuch@ericsson.com>