Remove shell Healtcheck test case
The healthcheck tier has grown thanks to the
adoption of SNAPs library. Now we have a SNAPs test
case that checks that creating a Network, a VM
and making sure it gets an IP from the DHCP agent
works. Basically, we are duplicating both test cases.

This patch removes the execution of the former test
without eliminating the code.

Change-Id: Iad6fec10fa1c247151c68313dc3760a55e511555
Signed-off-by: jose.lausuch <jose.lausuch@ericsson.com>