Update patch set 1
Patch Set 1:

> I would perfer 'pip install' to setup.py as Functest shipped in the
 > docker container is a module to be used instead of a tool to be
 > developed.
 > 
 > You can use 'pip install -U ~/repos/functest' to apply any update
 > to the module during development.

Sure, but people do develop it. If the purpose of the container is to be something that is shipped and not modified, the repos/ dir should be removed. The problem I've had with this* is that it feels quite natural/intuitive to modify things in the repos/ folder, try to run them and then you get weird messages about missing modules, since those modifications are not applied to the system version of the modules. 

*No me doing it per se, but I've had people repeatedly ask me why they changes are not working.

Patch-set: 1