Update patch set 1
Patch Set 1: -Code-Review

Any update on Domino Ulas?
for Colorado running Domino here was triggering Domino
but within Domino the test processing was skipped because some manual operations were required
Can you give a status?
if it is still the case, I am not sure it makes sense to reenable the triggering of the test here

Patch-set: 1
Label: Code-Review=0