Bugfix: make security_scan authorized
JIRA: FUNCTEST-732

"OS_AUTH_URL" in the file "stackrc" in undercloud is needed when
trying to obtain controller and compute nodes of openstack.
Thus, when running the security_scan.py, source stackrc is needed.
This patch will make security_scan work in CI until the patch in
securityscanning repo is merged:
https://gerrit.opnfv.org/gerrit/#/c/29705/

Change-Id: I379c3a71522973887d1ea698ab1df761354e6c36
Signed-off-by: Linda Wang <wangwulin@huawei.com>