Update patch set 1
Patch Set 1: Code-Review-1

I would actually prefer that by default is not blocking. Why? If I don't know about this behavior and I add a new test case, I will miss that flag, and if by default it's blocking, I might not be wanting that.

For cleaning, default True is ok. Although we will remove this in cleaning mechanism in E-release.

Patch-set: 1
Label: Code-Review=-1