Update patch set 2
Patch Set 2:

> (1 comment)
 > 
 > I don't understand why you use bash calls. Please let me know if I
 > miss something.

It is inherited from Feature_base, which is providing an abstraction for feature project testing. 
Most of them were running bash commands, that is why the abstraction expects a cmd.
security_scanning is a feature project BUT it is written in python....
as we considered it as a feature project (which is the case, moreover it is no more in functest repo since few weeks), we manage it the same way than other feature projects...

So two options
- python way: do not use Featurebase but Testbase but then we will miss some commons related to feature projects introduced in Featurebase

- feature way: accept the bash calls as an exception for this release (we are not supposed to write new code but just fix bugs..) and work on featurebase evolution to support also python calls

Patch-set: 2